backend/requirements.txt
fastapi
uvicorn
pymongo
python-multipart
pypdf
sentence-transformers
ollama
pydantic
python-dotenv
numpy
pydantic[email]
PyJWT
bcrypt
